refactor: 기존 나이 원시 값을 신규 나이 속성으로 리펙터링
