+++ src/main/java/nextstep/subway/member/dto/MemberRequest.java
@@ -1,5 +1,6 @@
 package nextstep.subway.member.dto;
 
+import nextstep.subway.member.domain.Age;
 import nextstep.subway.member.domain.Member;
 
 public class MemberRequest {
@@ -29,6 +30,6 @@ public class MemberRequest {
     }
 
     public Member toMember() {
-        return new Member(email, password, age);
+        return new Member(email, password, Age.from(age));
     }
 }
