+++ src/main/java/nextstep/subway/member/dto/MemberResponse.java
@@ -17,7 +17,7 @@ public class MemberResponse {
     }
 
     public static MemberResponse of(Member member) {
-        return new MemberResponse(member.getId(), member.getEmail(), member.getAge());
+        return new MemberResponse(member.getId(), member.getEmail(), member.getAgeIntValue());
     }
 
     public Long getId() {
